imx6ul: add missing uart DMA channels
This will also be included by IMX6ULL as it uses imx6ul.dtsi as a base.
Extra log added so we can quickly check in dmesg that DMA is used for UART ports.

Event number can be found in Reference Manuall page 189. UART6 does not support TX DMA so configuration for it is not included, as current driver requires both DMA channels to be present.